Renamed cli module to gui module
